Plug throught cloud-provider-name to usercluster controller
