+++ api/pkg/controller/usercluster/reconciler.go
@@ -352,8 +352,7 @@ func (r *reconciler) reconcileUnstructured(ctx context.Context) error {
 	}
 
 	creators := []reconciling.NamedUnstructuredCreatorGetter{
-		//openshift.InfrastructureCreatorGetter(r.platform),
-		openshift.InfrastructureCreatorGetter("AWS"),
+		openshift.InfrastructureCreatorGetter(r.platform),
 	}
 	r.log.Debug("Reconciling unstructured")
 	// The delegatingReader from the `mgr` always redirects request for unstructured.Unstructured
